Fix up onionhoney's code to work on real solves.
The analyzer code assumed that the user could not make
slice moves or cube rotations, and couldn't find any
split points in the presence of those moves. Make all
the solution finding check all orientations so that
it now mostly works. It probably still fails if a
double move crosses a boundary or in other edge
cases.

Add a wallaby.cjs file to deal with the slowness and size
of some of the onionhoney tests.