Fix missing moves and parsing in onionhoney code.
A few bugs were making solve breakdowns often not work.

Wide moves in three axes were missing; added them in.

E slice moves were coded backwards (aligned with U instead
of D). Fixed this and all moves that depended on the wrong-way
E slice.

Parsing couldn't handle R3 or greater or similar examples, or
things like R5' or L3'. Hacked it up so that it does recognize
these and preserve low movecount instances (like R3). This is
still not really right but nobody is going to do a double digit
turn during a speedsolve, so it might be a long while before it
needs fixing.

Add some basic tests for these new cases.