Found an additional premove reference.
When I excised premove in the previous change, missed a spot
that relied on it from page.svelte. Sloppy typing should really
be fixed, but for now just fix up the actual problem.

While I was at it, I added the unit test from last change in
with the correct reconstruction and ran a format.