Validate solve analyzer from almost every angle.
Discovered that the magical `inv()` call on orientation was largely
a no-op and so seemed right but basically never was; the orientation
is the transform that needs to be done to the cube to bring it from
WG to the orientation that is being used by the solver and by the
masks for the different stages. So it should be applied _without_
inverting it; but often, the orientation is its own inverse so the
fact that it was inconsistent and wrong was rarely showing up as
an issue.

Once I started trying to sort out those cases where it _wasn't_
working, a lengthy debugging session resulted in the money part of
this change: removal of calls to inv() on orientation almost
everywhere.

While I was at it I added test cases for roux solves from almost
every angle; all x2y options, and one random cn option. In doing
so I discovered that only when the user did orange/yellow there was
a new problem, and it was because the view push back code relied
on their being an orientation+view, which there isn't in this case.

At long last, it looks like the analyzer and optimizer can handle
every angle and give the user good feedback on their solve.