Cut down method reconstruction to just one case.
Instead of solving cases with cube rotations provided one way
and other cases a separate way, unify the codepath into a single
method that recognizes that views only apply if a cube rotation
was done and generates the same answers for the breakdown as
before.

Change in stages and reconstructions.unit.ts are due to formatting
only.