Add free and frozen faces to generic reconstructions.
Now the method can specify a face to be frozen in order to force a
wide move interpretation from that stage forward. Very useful for
Roux to force r and r' moves after fb, possibly useful for other
methods in the future. In this initial rather hacky implementation,
reconstruction only handles freezing the L face, but this will get
generalized in an upcoming change to handle freezing any face to
create the corresponding opposite wide move. Frozen faces apply
to every stage after the stage in which they are frozen.

Also, the method can specify a 'free face' for a stage match. This
enables cmll to match whether the AUF is done or not, and will also
make it easy to encode mismatched blocks for roux (e.g.  yellow on
the bottom instead of white on one side). This is implemented
properly and works to enable any face move to be free. Free faces
apply only to the stage they are provided for, and if you mean for
the face to remain free subsequent stages have to continue to
declare the fact.