Eliminate a factor of almost 3 million in auto pruner.
The automatic pruner was treating the centers as significant which
is needless because we can always just hold the cube so that the
centers are fully solved before assigning a coordinate based on the
positions of the pieces which do in fact move unlike the centers
which do not in fact move.

This gets rid of a factor of up to 2^6 * 6^6 in the pruning table,
or almost 3 million in the worst case.