Fix up solver to be properly configurable.
Make the solver take in movesets for each edge in the DAG, and
set the pruner size based on the size of the moveset to enable
actually solving CMLL in a reasonable timeframe.

Change the method editor to have a more fool-proof cn setting for
the first phase, and to let the user specify the movesets for
each edge.

Pruning tables can now result from searches of size up to
20 million for more efficient and effective solving, especially
for CMLL and LSE stages.

This still needs to get shoved into a worker because it can cause
multi-second delays, but it is fairly usable now.