Fix a fundamental encoder bug.
I had convinced myself that by putting the centers in canonical
positions I could take advantage of symmetries of the cube in
the encoder, but this is wrong because the encoding will then
prune non-solved states to 'solved' by encoding them the same.

There ought to be some way to do this bit of cleverness but for
now just make it correct not clever.