Rejected splitting moves for solve analysis.
I think it will turn out to be better to make the move sequence
match as closely as possible the actual turns on the bluetooth
cube, and not do any splitting here to cause the processing to
recognize stages early.

The user can add free faces to their solve states if there is
a transition that is passed through with a move like R2, and
I don't want to find splits that don't correspond to cube
states the user never saw.