Split out cube code into a new TCube component.
Start to turn the prototype into more of a real cube player by
making it its own component.

Push the essentially wrong-headed "just use props to track the
rotation" approach about as far as it goes. Map the rotations
onto the correct cubies for the move, and move them.

The essential issue being that this way of transforming doesn't
stack in a logical way. Everything keeps rotating according to
the world axes they started with, instead of rotating around
the real world x, y, or z; so the first rotation is fine, but
the second one is in the wrong axis. Other than that it works
great :-)

Accumulating a quaternion or transform matrix per object should
be a much more sensible way to accomplish the same thing, and
have the rotations compose as expected.