Rotations animated.
An almost working version of animated rotations which relies on
three.js to slerp the rotation from one quaternion to the next.
The problem is it doesn't always pick the same way to get there and
so the cubies don't necessarily move together in a nice block, so
some sort of manual interpolation is going to be needed instead.