"Fully working" animations of moves and tidying of player.
The new player is almost ready for prime time with properly
animating moves (when they overlap it gets a little funky but
the effect isn't bad) that never leave the cubies out of sync
or out of place.

The algorithm viewer now fits nicely in the player and can't
be rewound or fast forwarded out of range.

Overall it's ready to be turned into the 'maximum bandage' cube.